// web/src/types/themeTypes/ThemeType.ts
export type ThemeTypes = {
  name: string;
  dark: boolean;
  variables?: object;
  colors: {
    primary?: string;
    primary20?: string;
    primary30?: string;
    primary50?: string;
    primary40?: string;
    primary60?: string;
    secondary?: string;
    secondary20?: string;
    secondary30?: string;
    secondary50?: string;
    secondary60?: string;
    grey?: string;
    greyDark?: string;
    greyLight?: string;
    info?: string;
    success?: string;
    accent?: string;
    warning?: string;
    error?: string;
    textPrimary?: string;
    textSecondary?: string;
    textSecondaryLight?: string;
    textDisable?: string;
    textDisable1?: string;
    borderColor?: string;
    hoverColor?: string;
    inputBorder?: string;
    containerBg?: string;
    background?: string;
    surface?: string;
    white?: string;
    black?: string;
    'on-surface': string;
    'on-surface-variant'?: string;
    bg?: string;
  };
};